Refactor extract tests into separate module, use maven plugin to enhance tests
So the motivation here is to move away from the test code enhancing classes in place. The issue with that is that depending on the file system the old un-enhanced classes are read and so the tests only pass on a second run (we have to run the tests twice to pass with the second NOT using clean).