Support transient initialisation in default constructor for other types (non-persistent ones)
The code that captures initialisation code was originally designed to capture and consume
initialisation of OneToMany and ManyToMany (thus ArrayList, LinkedHashSet, HashSet types).

This change defers the check of the type being initialised for the OneToMany,ManyToMany,DbArray
cases and thus allows the use of this capturing for transient fields being initialised. We
want this when we have entity classes without default constructors ... and for this case when
ebean enhancement adds the default constructor we want to also add in code that initialises
the transient fields by replaying the code captured when visiting the constructor(s).