Add error detection for transient field initialisation in default constructor
When ebean-agent is adding a default constructor and there are transient fields that
it can not initialise it will provide a nice error message or throw based on
transient-init and transient-init-error

If a transient field is initialised in a constructor it must be a simple <init> using
a no argument default constructor of a type. Initialisation via any other means is
not supported (using a method or non-default constructors are not supported).

When this limitation is hit folks can opt to add a default constructor themselves.
They get full control and ebean-agent does not need to add a default constructor.

Note that "Not Supported" in effect means that the ebean-agent is not be able to
initialise those transient fields in the default constructor that it is adding.
Those fields will be uninitialised so null, 0, false etc when ebean creates
the instance [using this default constructor].