#229 AntEnhanceTask to honor packages specified
This means that packages specified will be honored over any
filtering that would otherwise occur via IgnoreClassHelper.

The fix is to combine the Ant task packages parameter with
the transform args and pass that in ultimately to the
EnhanceContext (which parses that out and will then use
the PackageFilter rather than IgnoreClassHelper).